Align server chunk send radius with client eviction (terrain_chunk_radius),
eliminating thousands of wasted chunks at elevation. Remove summary_only
flag, simplify VisibleChunkCache. Decorator.index narrowed usize→u16.

Network transport wrapper (ADR-034) replacing direct bevy_renet access.
Server-authoritative chunk eviction via EvictChunks message. Fix
MessageReader/Writer<Do> conflict in do_incremental.

Metrics overlay: seg_row abstraction with auto gaps, composite segments
split into halves, colocated format constants replacing shared presets.
Frame p95/FPS cached at sample interval. Console: overrun alarm coloring,
per-channel queue/buffer stats.

+++ crates/common-bevy/src/components/entity_type/decorator.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Copy, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
 pub struct Decorator {
-    pub index: usize,
+    pub index: u16,
     pub is_solid: bool,
 }
